+++ items/odd_chars/diacritics_k.sh
@@ -41,6 +41,7 @@ no_auto_menu_handling=1 # delay processing of dialog, only source it for now
 if [ -n "$1" ]; then
     "$D_TM_BASE_PATH"/scripts/act_display_char.sh "$1"
 elif $b_use_alt_handler; then
+    ${all_helpers_sourced:-false} || source_all_helpers "diacritics_k.sh"
     tmux_error_handler set-option -gqu "$wt_pasting"
 fi